chore(): improve user management
