+++ functions/modules/user.module.js
@@ -16,17 +16,18 @@ const fetchAll = async (req, res) => {
 }
 
 const fetchOne = async (req, res) => {
-  const userId = req.param('id');
+  const userId = req.params.id;
   try {
     const userRecord = await admin.auth().getUser(userId);
     res.send({ data: userRecord.toJSON() });
   } catch (error) {
+    res.status(400);
     return res.send(error);
   }
 }
 
 const update = async (req, res) => {
-  const userId = req.param('id');
+  const userId = req.params.id;
   const { data } = req.body;
 
   try {
@@ -38,8 +39,21 @@ const update = async (req, res) => {
   }
 }
 
+const create = async (req, res) => {
+  const { data } = req.body;
+  try {
+    const userRecord = await admin.auth().createUser({ ... data, emailVerified: false, disabled: false });
+    // await admin.auth().generateEmailVerificationLink(data.email);
+    res.send({ data: userRecord.toJSON() });
+  } catch (error) {
+    res.status(400);
+    res.send(error);
+  }
+}
+
 
 router.post('/', fetchAll);
+router.post('/create', create);
 router.post('/:id', fetchOne);
 router.post('/:id/update', update);
