+++ src/common/helpers/extract-permissions-from-roles.js
@@ -1,0 +1,4 @@
+import uniq from 'lodash/uniq';
+
+export default roles => uniq(roles.map(role => role.permissions)
+  .reduce((prev, current) => [...prev, ...current], []));
