+++ src/components/User/UserForm/UserForm.md
@@ -1,4 +1,3 @@
 # UserForm
 
-A simple user form with validation enabled.
-It is possible to intialize the form by passing values though props.
+User form with User identity form and roles management.
