+++ src/components/User/UserIdentityForm/UserIdentityForm.stories.js
@@ -1,0 +1,18 @@
+/* eslint-disable import/first */
+/* eslint-disable import/no-extraneous-dependencies */
+import { storiesOf } from '@storybook/vue';
+import { action } from '@storybook/addon-actions';
+import UserIdentityForm from '.';
+import markdown from './UserIdentityForm.md';
+
+export default {
+  title: 'UserIdentityForm',
+};
+
+
+storiesOf('UserIdentityForm', module)
+  .add('default', () => ({
+    components: { UserIdentityForm },
+    template: '<UserIdentityForm @submit="action" display-name="john" />',
+    methods: { action: action('submit') },
+  }), { notes: { markdown } });
