+++ src/components/User/UserIdentityForm/index.js
@@ -1,0 +1,1 @@
+export { default } from './UserIdentityForm.vue';
