+++ src/views/Dashboard.vue
@@ -24,14 +24,29 @@
             <v-list-item-title>Roles</v-list-item-title>
           </v-list-item-content>
         </v-list-item>
-        <v-list-item :to="{ name: 'users' }">
-          <v-list-item-action>
-            <v-icon>mdi-account-supervisor</v-icon>
-          </v-list-item-action>
-          <v-list-item-content>
+        <v-list-group
+          prepend-icon="account_circle"
+        >
+          <template v-slot:activator>
             <v-list-item-title>Users</v-list-item-title>
-          </v-list-item-content>
-        </v-list-item>
+          </template>
+          <v-list-item :to="{ name: 'users' }">
+            <v-list-item-action>
+              <v-icon>supervised_user_circle</v-icon>
+            </v-list-item-action>
+            <v-list-item-content>
+              <v-list-item-title>Manage users</v-list-item-title>
+            </v-list-item-content>
+          </v-list-item>
+          <v-list-item :to="{ name: 'newUser' }">
+            <v-list-item-action>
+              <v-icon>person_add</v-icon>
+            </v-list-item-action>
+            <v-list-item-content>
+              <v-list-item-title>New User</v-list-item-title>
+            </v-list-item-content>
+          </v-list-item>
+        </v-list-group>
         <v-list-item>
           <v-list-item-action>
             <v-icon>mdi-calendar-clock</v-icon>
